Deprecated --cxinstalldir and --cxginstalldir
The path to a crossover installation can now be provided directly as an
argument for --wine.

This simplifies a lot of the logic for handling various CX versions, and it
makes it easier to switch to alternate CX versions by supplying --from and --wine
together (ie. one may now use "--from script --wine cx" to use the system default
cx installation, where previously that would have obeyed Cxinstalldir from the
script, potentionally not changing anything).

This bumps the meta format to v4.